Add basic view into miditable
This commit adds most of what is needed to have a visual representation of the
midi connections for a small program, though under ntk damage() does not seem to
result in the expected redraws so things are a bit glitchy.
Port registrations are read from the midi table and displayed in a simple
alphabetical format.
For the large array of controls that are possible in zyn it might make sense for
each window to know what controls it has in it and just display those in a per
window mapping interface, though this approach will add more windows to an
already complicated setup...