rtosc.h: Add Argument Iterator
Previously accessing all N arguments in a message was an O(N^2) operation
with the use of an iterator this cost is reduced to O(N).
This enables easier handling of messages with arguments like [f*] which
were previously avoided in favor of blob arguments.

Three new methods are exported from the C layer:

    rtosc_arg_itr_t rtosc_itr_begin(const char *msg);
    rtosc_arg_val_t rtosc_itr_next(rtosc_arg_itr_t *itr);
    int rtosc_itr_end(rtosc_arg_itr_t itr);