Experiment With Extending typed-message.h
This incorporates http://github.com/irrequietus/typestring
which if I've understood the license is designed for this sort of
application.
Ideally typed messages recognize multiple levels of OSC paths in a reasonably
efficient manner.
This when this feature is finished, path matching in a quasi-compatible manner
with respect to OSC 1.0 should be implemented.

This code is incomplete, but the completed portion has tests in:
- test/patterns.c
- test/typed-template-test.cpp