path-search: Document ambiguity for invalid ports
Document ambiguities of path-search in test/path-search.cpp.

If there is a port where there was a trailing slash without
`Port::ports` (no subports), path-search of a query "/x/"
would return "x/", making the querier assume that there was
a sub-port x/ inside "/x/". The querier was continuing with
path-search on "/x/x/", which returned "x/" again. Then it
was querying "/x/x/x"...

A query in port-checker which resulted in an infinite query
in further commits is being disabled now. This disables the
query for "trailing slash without subports", which will,
however, be detectable by infinitely recursing apps.