Fix ThreadLink overwriting messages
This fixes an issue where a write to ThreadLink after a contigous write
would corrupt the last byte of the previous message.

Before the patch, when ThreadLink would want to split the buffer into 2
halves, the sizes were computed as

```
const size_t w1 = ring->size - ring->write - 1
const size_t w2 = len - w1
```

This was wasting one (usable) byte - the last byte of `ring->buffer`.
However, the real problem with this was that the subsequent write
position was *not* wasting this byte:

```
const off_t  next_write = (ring->write + len)%ring->size
```

So these 2 variables did not fit each other, and the next write position
was calculated one byte too low.

The bugfix is to simply turn off the behavior of wasting the last byte
of `ring->buffer` in `ring_write`. The read functions are patched
accordingly.

A test has been added that fails without this patch.