thread_link: Add read lookahead
Add read lookhead feature, whereby read operations can be set to read from
a lookahead queue, which is the same as the ordinary read queue,
except the messages are not considered removed from the ring, but
remain for a future non-lookahead read to actually consume.

The lookahead queue is synchronized to the ordinary read queue
whenever an ordinary read is performed.