pretty-format: Fix crash reading some ranges
Some ranges like

`5 5 ... 10`

caused the parser to abort (`if(!cmp)` in `delta_from_arg_vals`).

The parser complained that the `llhsarg` and `lhsarg` (in this case both
`5`) are equal, and therefore, a step size can not be inferred. However,
in this situation, the step size shall be defined as implicitly +1 or
-1.

The printer prints ranges in that way intentionally, and this was
also the idea behind it, so the parser should read it, too.