Savefile: Build port graph to fix dependencies
Before this commit, the messages are dispatched in two rounds, first the
ports on which other ports depend, then the other ports. However, the
complexity can be higher, so this commit first creates a graph of the
messages that shall be dispatched, then draws dependency edges, and then
dispatches the messages in topological order.