Fixup: Re-add `-fPIC`
Fixup of c00fdb161004e06e52ef46ba921014bd15b1601b

The mentioned commit removes the `-fPIC` flag, but this is needed in
Zyn: Even though zyn links rtosc as a `.a` file, the plugin versions are
shared objects and do require that every part, including rtosc is
compiled with `-fPIC`.

On the other side it is desired to compile rtosc as a `.a` into zyn to
be able to use link time optimization (LTO) and for distributional
reasons.