Adjust tests for liblo-0.32
This adjusts the `performance` test due to liblo's commit

4899a0d9653a347ee25cf5013ebddc615b1ef80f

The rtosc code was requested to stay as it is - i.e. to still allow
messages not starting with `/`.

Also, this changes Windows build to take Release, which gets around the
(still existing) liblo issue where the Debug execution of `performance`
hits the assertion

```
Assertion `d.matches == 3600000' failed.
```

(which causes infinite test execution on Windows due to failed assert).
Using Windows `Release` build allows us to run the `performance` test
with liblo 0.32.